Se realiza reserva pero se deja pendiente creacion de usuarioentity

+++ ms3/src/main/java/com/tingeso/ms3/Entity/DiscountFrecuenciaEntity.java
@@ -16,5 +16,5 @@ public class DiscountFrecuenciaEntity {
     private String categoria;
     private int minVisitas;
     private int maxVisitas;
-    private int descuento;
+    private double descuento;
 }
